+++ CHANGES.rst
@@ -18,6 +18,7 @@ general
 lqdn-gafam-campaign
 -------------------
 - Enable RTL and adjust layout for translation into Arabic
+- Use English translation of *colored* summary image as og:image and twitter:image in chooser.html
 
 
 2018-05-05 0.5.0